Initial commit of class for computing mortar segment mesh elements in 3D.
Currently the class computes and outputs the polygon clipping and triangulation of secondary elements against each primary element individually. However, this does not allow for sections of secondary elements without a corresponding primary element to be triangulated. This class is therefore currently restricted to problems where the primary interface face projection covers the secondary interface.